I was having an issue with the bx-loader div displaying I found it easier to append the class to bx-wrapper and style the bx-wrapper.bx-loader selection chain
the jquery.bxslider.css file will need adjustment to compensate for the  new, different markup. Instead of .bx-wrapper .bx-loading being the selector to display the loading .gif, we now need to target .bx-wrapper.bx-loading 

I did not modify the css file as I am uncertain as to how well these changes fit into the overall direction of bxslider. My implementation strays quite a bit from the default setup and I am unsure whether my issue with the bx-loading div is isolated to my usage or is an actual bug.